fix(deleteExercise): Allow the admin to delete any item from the database
Connected the frontend buttons to delete items to the backend routes. Created validation requirements for the deletion of exercises, notes and news. Modified the Tag table to establish a relationship with Excercise and Note

Currently, a user account cannot be deleted

+++ icpc-backend/src/categories/categories.service.ts
@@ -5,6 +5,13 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Category } from './entities/category.entity';
 import { Comment } from '../comment/entities/comment.entity';
+import {
+  Ticket,
+  TicketOperation,
+  TicketStatus
+} from 'src/ticket/entities/ticket.entity';
+import { Note } from 'src/notes/entities/note.entity';
+import { Excercise } from 'src/excercises/entities/excercise.entity';
 
 @Injectable()
 export class CategoriesService {
@@ -12,7 +19,13 @@ export class CategoriesService {
     @InjectRepository(Category)
     private readonly categoryRepository: Repository<Category>,
     @InjectRepository(Comment)
-    private readonly commentRepository: Repository<Comment>
+    private readonly commentRepository: Repository<Comment>,
+    @InjectRepository(Ticket)
+    private readonly ticketRepository: Repository<Ticket>,
+    @InjectRepository(Note)
+    private readonly notesRepository: Repository<Note>,
+    @InjectRepository(Excercise)
+    private readonly excerciseRepository: Repository<Excercise>
   ) {}
   async create(createCategoryDto: CreateCategoryDto) {
     const name = await this.findOneByName(createCategoryDto.name); // check if name exists
@@ -69,7 +82,48 @@ export class CategoriesService {
   }
 
   async remove(id: string) {
-    const category = await this.findOne(id);
-    return await this.categoryRepository.remove(category);
+    const category = await this.categoryRepository
+      .createQueryBuilder('category')
+      .leftJoinAndSelect('category.excercises', 'excercises')
+      .leftJoinAndSelect('category.notes', 'notes')
+      .where('category.id = :id', { id })
+      .getOne();
+    const ticketCommentBody = `La categoría ${category.name} ha sido eliminada`;
+    const ticketComment = this.commentRepository.create({
+      body: ticketCommentBody
+    });
+    const ticketCommentId = await this.commentRepository.save(ticketComment);
+    const ticket = this.ticketRepository.create({
+      operation: TicketOperation.DELETE,
+      status: TicketStatus.ACCEPTED,
+      commentId: ticketCommentId
+    });
+    const savedTicket = await this.ticketRepository.save(ticket);
+    const categories = await this.categoryRepository.find({});
+    let pivot = categories[0];
+    if (pivot.id === category.id) {
+      pivot = categories[1];
+    }
+    if (savedTicket) {
+      if (categories.length === 1) {
+        throw new BadRequestException(
+          'No se puede eliminar la única categoría'
+        );
+      } else {
+        if (category.excercises.length > 0) {
+          for (const exercise of category.excercises) {
+            exercise.category = pivot;
+            await this.excerciseRepository.save(exercise);
+          }
+        }
+        if (category.notes.length > 0) {
+          for (const note of category.notes) {
+            note.category = pivot;
+            await this.notesRepository.save(note);
+          }
+        }
+        return await this.categoryRepository.remove(category);
+      }
+    }
   }
 }
