+++ icpc-backend/src/memory/memory.service.ts
@@ -1,15 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateMemoryDto } from './dto/create-memory.dto';
 import { UpdateMemoryDto } from './dto/update-memory.dto';
 import { Memory } from './entities/memory.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { Comment } from 'src/comment/entities/comment.entity';
+import {
+  Ticket,
+  TicketOperation,
+  TicketStatus
+} from 'src/ticket/entities/ticket.entity';
+import { Excercise } from 'src/excercises/entities/excercise.entity';
 
 @Injectable()
 export class MemoryService {
   constructor(
     @InjectRepository(Memory)
-    private readonly memoryRepository: Repository<Memory>
+    private readonly memoryRepository: Repository<Memory>,
+    @InjectRepository(Comment)
+    private readonly commentRepository: Repository<Comment>,
+    @InjectRepository(Ticket)
+    private readonly ticketRepository: Repository<Ticket>,
+    @InjectRepository(Excercise)
+    private readonly excerciseRepository: Repository<Excercise>
   ) {}
 
   async create(createMemoryDto: CreateMemoryDto) {
@@ -50,7 +63,40 @@ export class MemoryService {
   }
 
   async remove(id: string) {
-    const memory = await this.memoryRepository.findOneBy({ id });
-    return await this.memoryRepository.remove(memory);
+    const allMemoryLimits = await this.memoryRepository.find();
+    if (allMemoryLimits.length === 1) {
+      throw new BadRequestException(
+        'No se puede eliminar el único límite de memoria'
+      );
+    }
+    let pivot = allMemoryLimits[0];
+    if (pivot.id === id) {
+      pivot = allMemoryLimits[1];
+    }
+    const memory = await this.memoryRepository
+      .createQueryBuilder('memory')
+      .where('memory.id = :id', { id })
+      .leftJoinAndSelect('memory.excercises', 'excercises')
+      .getOne();
+    const ticketCommentBody = `El límite de memoria ${memory.memoryLimit.toString()} ha sido eliminado`;
+    const comment = this.commentRepository.create({ body: ticketCommentBody });
+    const savedComment = await this.commentRepository.save(comment);
+    const ticket = this.ticketRepository.create({
+      operation: TicketOperation.DELETE,
+      status: TicketStatus.ACCEPTED,
+      commentId: savedComment
+    });
+    const savedTicket = await this.ticketRepository.save(ticket);
+    if (savedTicket) {
+      if (memory.excercises.length > 0) {
+        for (const exercise of memory.excercises) {
+          exercise.memoryId = pivot;
+          await this.excerciseRepository.save(exercise);
+        }
+      }
+      return await this.memoryRepository.remove(memory);
+    } else {
+      return null;
+    }
   }
 }
